% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp_eventPred.R
\name{runShinyApp_eventPred}
\alias{runShinyApp_eventPred}
\title{Run Shiny app}
\usage{
runShinyApp_eventPred()
}
\description{
Runs the event prediction Shiny app.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
