% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEnrollment.R
\name{predictEnrollment}
\alias{predictEnrollment}
\title{Predict enrollment}
\usage{
predictEnrollment(
  df = NULL,
  target_n,
  enroll_fit,
  lags = 30,
  pilevel = 0.9,
  nyears = 4,
  nreps = 500,
  showsummary = TRUE,
  showplot = TRUE
)
}
\arguments{
\item{df}{The subject-level enrollment data, including \code{trialsdt},
\code{randdt} and \code{cutoffdt}. By default, it is set to
\code{NULL} for enrollment prediction at the design stage.}

\item{target_n}{The target number of subjects to enroll in the study.}

\item{enroll_fit}{The pre-fitted enrollment model used to
generate predictions.}

\item{lags}{The day lags to compute the average enrollment rate to
carry forward for the B-spline enrollment model. By default,
it is set to 30.}

\item{pilevel}{The prediction interval level. By default,
it is set to 0.90.}

\item{nyears}{The number of years after the data cut for prediction.
By default, it is set to 4.}

\item{nreps}{The number of replications for simulation. By default,
it is set to 500.}

\item{showsummary}{A Boolean variable to control whether or not to
show the prediction summary. By default, it is set to \code{TRUE}.}

\item{showplot}{A Boolean variable to control whether or not to
show the prediction plot. By default, it is set to \code{TRUE}.}
}
\value{
A list of prediction results, which includes important information
such as the median, lower and upper percentiles for the estimated
time to reach the target number of subjects, as well as simulated
enrollment data for new subjects. The data for the
prediction plot is also included within the list.
}
\description{
Utilizes a pre-fitted enrollment model to generate
enrollment times for new subjects and provide a prediction
interval for the expected time to reach the enrollment target.
}
\details{
The \code{enroll_fit} variable can be used for enrollment prediction
at the design stage. A piecewise Poisson model can be parameterized
through the time intervals, \code{accrualTime}, which is
treated as fixed, and the enrollment rates in the intervals,
\code{accrualIntensity}, the log of which is used as the
model parameter. For the homogeneous Poisson, time-decay,
and piecewise Poisson models, \code{enroll_fit} is used to
specify the prior distribution of model parameters, with
a very small variance being used to fix the parameter values.
It should be noted that the B-spline model is not appropriate
for use during the design stage.

During the enrollment stage, \code{enroll_fit} is the enrollment model
fit based on the observed data. The fitted enrollment model is used to
generate enrollment times for new subjects.
}
\examples{

# Enrollment prediction at the design stage

enroll_pred <- predictEnrollment(
  target_n = 300,
  enroll_fit = list(model = "piecewise poisson",
                    theta = log(26/9*seq(1, 9)/30.4375),
                    vtheta = diag(9)*1e-8,
                    accrualTime = seq(0, 8)*30.4375),
  pilevel = 0.90, nreps = 100)


}
